hw3-http-client/http-handler/requirements.txt

functions-framework==3.4.0
google-cloud-storage==2.11.0
google-cloud-pubsub==2.18.4